[debug] move nf4 op registration from utils to modules
