+++ torchtune/modules/peft/lora.py
@@ -11,8 +11,8 @@ import torch.nn.functional as F
 from torch import nn, Tensor
 
 from torchao.dtypes.nf4tensor import linear_nf4, to_nf4
+from torchtune.modules.low_precision import _register_nf4_dispatch_ops  # noqa: F401
 from torchtune.modules.peft.peft_utils import AdapterModule
-from torchtune.utils import _register_nf4_dispatch_ops  # noqa: F401
 
 
 class LoRALinear(nn.Module, AdapterModule):
